Add files via upload
A/B tests are very commonly performed by data analysts and data scientists.
For this project, A/B test is ran by an e-commerce website. The goal is to work through this notebook to help the company understand if they should implement the new page, keep the old page, or perhaps run the experiment longer to make their decision.