Add files via upload
In this project, An investigation of a classic phenomenon from experimental psychology called the Stroop Effect will be done.The aim is to learn a little bit about the experiment, create a hypothesis regarding the outcome of the task, then go through the task. Then some data collected be investigated for subjects who have performed the same task/experiment and will compute some statistics describing the results. Finally, the results will be interpreted in terms of your hypotheses and T-Test.